chore: migrate to vitest
