--- jest.config.base.js
@@ -1,11 +1,0 @@
-export default {
-  preset: 'ts-jest',
-  testEnvironment: 'jsdom',
-  coverageThreshold: {
-    global: {
-      branches: 5,
-      functions: 5,
-      lines: 5,
-    },
-  },
-};
