+++ package.json
@@ -9,6 +9,7 @@
     "clean:install": "npm run clean && rimraf node_modules .nx package-lock.json && npm i",
     "test": "npx nx run-many -t test",
     "test:coverage": "npx nx run-many -t test:coverage",
+    "test:browser": "npx nx run-many -t test:browser",
     "lint": "npx nx run-many -t lint",
     "build": "npx nx run-many -t build",
     "pack": "npx nx run-many -t pack",
@@ -29,28 +30,30 @@
     "@rollup/plugin-terser": "^0.4.4",
     "@rollup/plugin-typescript": "^11.1.6",
     "@types/eslint__js": "^8.42.3",
-    "@types/jest": "^29.5.8",
+    "@vitest/browser": "^2.0.5",
+    "@vitest/coverage-v8": "^2.0.5",
+    "@vitest/ui": "^2.0.5",
     "bundlesize": "^0.18.2",
     "eslint": "^9.9.1",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-jest": "^28.8.2",
     "eslint-plugin-jsdoc": "^50.2.2",
     "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-vitest": "^0.5.4",
     "globals": "^15.9.0",
     "husky": "^9.1.5",
-    "jest": "^29.7.0",
-    "jest-environment-jsdom": "^29.7.0",
+    "jsdom": "^25.0.0",
     "lint-staged": "^15.2.10",
     "nx": "19.6.5",
+    "playwright": "^1.46.1",
     "prettier": "^3.0.3",
     "rimraf": "^6.0.1",
     "rollup": "^4.21.2",
     "rollup-plugin-copy": "^3.5.0",
     "rollup-plugin-dts": "^6.1.1",
     "rollup-plugin-visualizer": "^5.12.0",
-    "ts-jest": "^29.1.1",
     "typedoc": "^0.26.6",
     "typescript": "~5.5.4",
-    "typescript-eslint": "^8.4.0"
+    "typescript-eslint": "^8.4.0",
+    "vitest": "^2.0.5"
   }
 }
