+++ packages/hls-parser/package.json
@@ -8,8 +8,10 @@
     "build": "npm run clean && rollup -c rollup.config.js && npm run build:api-reference",
     "build:api-reference": "npx typedoc --options typedoc.json",
     "verify:bundle-size": "npx bundlesize --config bundlesize.json",
-    "test": "jest",
-    "test:coverage": "jest --coverage=true",
+    "test": "vitest run",
+    "test:coverage": "vitest run --coverage",
+    "test:vitest-ui": "vitest --ui",
+    "test:browser": "npx playwright install && vitest run --browser",
     "lint": "eslint .",
     "pack": "npm pack"
   },
