--- packages/playback/jest.config.js
@@ -1,7 +1,0 @@
-import base from '../../jest.config.base.js';
-
-export default {
-  ...base,
-  displayName: 'playback',
-  coverageDirectory: './coverage',
-};
