+++ packages/playback/test/utils/timeRanges.test.ts
@@ -1,4 +1,4 @@
-import { describe, it, expect } from '@jest/globals';
+import { describe, it, expect } from 'vitest';
 import PlayerTimeRange from '../../src/lib/utils/timeRanges';
 
 class MockTimeRange implements TimeRanges {
