+++ vitest.config.base.ts
@@ -1,0 +1,12 @@
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+  test: {
+    environment: 'jsdom',
+    browser: {
+      provider: 'playwright',
+      name: 'chromium',
+      headless: true,
+    },
+  },
+});
